Drop 'generic' FreeRTOS things and make it ESP-IDF specific
Instead of having a generic implementation, make a dedicated
implementation for ESP-IDF. This makes better integration with things
like logging system possible. And allows the freertos version to focus
on a pure FreeRTOS implementation of it.
Regular updates from the FreeRTOS implementation to update API for
things like the mutex can still be manually applied if needed.

This change does mean that some of the modules are completely rewritten
such that they use the native API's from ESP-IDF instead of reinventing
the wheel.